fix minor warnings
