+++ src/app/services/setting.service.ts
@@ -1,15 +1,20 @@
 import { Injectable } from "@angular/core";
 import { BaseDirectory } from "@tauri-apps/api/path";
-import {
-  latLng,
-  LatLng,
-  LatLngBounds,
-  MarkerClusterGroupOptions,
-} from "leaflet";
+import { LatLngBounds, MarkerClusterGroupOptions } from "leaflet";
 import * as fs from "@tauri-apps/plugin-fs";
 
 export type MapSettings = {
   initialMapBounds?: LatLngBounds;
+  maxZoomLevel?: number;
+  markerClusterGroupOptions: MarkerClusterGroupOptions;
+  showLabels?: boolean;
+};
+
+type MapSettingsWritten = {
+  initialMapBounds?: {
+    _southWest: [number, number];
+    _northEast: [number, number];
+  };
   maxZoomLevel?: number;
   markerClusterGroupOptions: MarkerClusterGroupOptions;
   showLabels?: boolean;
@@ -41,7 +46,7 @@ export class SettingService {
         baseDir: BaseDirectory.AppData,
       })
       .then((content) => {
-        const settings: any = JSON.parse(content);
+        const settings: MapSettingsWritten = JSON.parse(content);
         return {
           ...settings,
           initialMapBounds: settings.initialMapBounds
